Added Album Loving & Database Update
Added ability to love albums. This required a database update. I opted to create a new Album table rather than add new columns in the Track table so that the Album has a shared data point among the tracks. Consider if using the Tracks table for Album data, then you favorite an album, the database gets updated for all tracks. But then, what if you fix an incorrect album tag on one file and it gets added to that album. For this file, the "AlbumLove" property isn't set, which could lead to Dopamine showing that the album isn't favorited if the user searches for that song (this is just one scenario).