Selective "None Selected"
When the Albums button is pressed (the Select None button) and using a custom sort method, only show albums from that custom sort method.

There seems to be an issue where you need to click an album before clicking the Albums button again. I hoped that changing the Album sort order would allow you to click the Albums button again, but no, you still need to click an album first. The effect of this is that albums that are not in the albums view are still in the tracks view until you click an album and click the Albums button again. Would like to "fix" this later.

Also of note, I did experiment with calling "await this.GetTracksAsync(null, null, SelectiveSelectedAlbums, this.TrackOrder);" inside ToggleAlbumOrderAsync(), but that led to a noticeable decrease in performance when changing album sort order.