Removed AlbumsHolder
Removed the previously added AlbumsHolder. It was added to keep a separate unfiltered list of Albums, but with better use of the CVS filtering feature, it is no longer necessary.

"Select none" is now clickable again after changing the album sort order, whereas before, you needed to click an album to be able to click it again.

A few bug fixes as well.